디자인 수정: 맨 위로 버튼 컴포넌트화 및 스타일 개선
