+++ src/pages/StoreDetailPage.jsx
@@ -6,6 +6,7 @@ import { Map, MapMarker } from 'react-kakao-maps-sdk' //카카오맵 추가
 import { getStoreById } from '../api/storeApi'
 import defaultImage from '../assets/images/luckeat-default.png'
 import bakerDefaultImage from '../assets/images/제빵사디폴트이미지.png'
+import ScrollTopButton from '../components/common/ScrollTopButton'
 
 function StoreDetailPage() {
   const { id } = useParams()
@@ -122,18 +123,7 @@ function StoreDetailPage() {
 
   // 스크롤 관련 함수
   const handleScroll = () => {
-    if (mainContainerRef.current) {
-      setShowScrollTopButton(mainContainerRef.current.scrollTop > 300)
-    }
-  }
-
-  const scrollToTop = () => {
-    if (mainContainerRef.current) {
-      mainContainerRef.current.scrollTo({
-        top: 0,
-        behavior: 'smooth',
-      })
-    }
+    // 스크롤 이벤트는 유지하되 맨 위로 버튼 관련 코드 제거
   }
 
   // 탭 클릭 시 해당 섹션으로 스크롤
@@ -629,29 +619,8 @@ function StoreDetailPage() {
           )}
         </div>
 
-        {/* 맨 위로 버튼 */}
-        {showScrollTopButton && (
-          <button
-            onClick={scrollToTop}
-            className="fixed bottom-24 left-1/2 transform -translate-x-1/2 translate-x-28 bg-yellow-500 text-white w-10 h-10 rounded-full flex items-center justify-center shadow-lg z-10 hover:bg-yellow-600"
-            aria-label="맨 위로 스크롤"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              className="h-6 w-6"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M5 10l7-7m0 0l7 7m-7-7v18"
-              />
-            </svg>
-          </button>
-        )}
+        {/* ScrollTopButton 컴포넌트 추가 */}
+        <ScrollTopButton />
       </div>
 
       {/* 전화번호 팝업 */}
